--- apps/go-test-caching/.eslintrc.json
@@ -1,13 +1,0 @@
-{
-  "extends": ["../../.eslintrc.json"],
-  "ignorePatterns": ["!**/*", "node_modules"],
-  "overrides": [
-    {
-      "files": ["./package.json"],
-      "parser": "jsonc-eslint-parser",
-      "rules": {
-        "@nrwl/nx/nx-plugin-checks": "error"
-      }
-    }
-  ]
-}
