--- apps/go-test-caching/CHANGELOG.md
@@ -1,1 +1,0 @@
-# go-test-caching
