--- apps/go-test-caching/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "go-test-caching",
-  "version": "0.0.0",
-  "private": true,
-  "description": "",
-  "engines": {
-    "pnpm": "9",
-    "node": "20"
-  },
-  "main": "dist/index.js",
-  "scripts": {
-    "local": "GITHUB_STEP_SUMMARY=/dev/null ts-node src/index.ts",
-    "test": "vitest"
-  },
-  "devDependencies": {
-    "@types/fs-extra": "^11.0.4",
-    "@types/node": "^20.14.9",
-    "@types/node-fetch": "^2.6.11",
-    "fs-extra": "^11.2.0",
-    "nock": "^13.5.4",
-    "node-fetch": "^2.7.0",
-    "simple-git": "^3.26.0",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.4.5"
-  },
-  "dependencies": {
-    "@actions/artifact": "^2.1.11",
-    "@actions/core": "^1.10.1",
-    "@actions/github": "^6.0.0",
-    "@actions/glob": "^0.4.0",
-    "@octokit/plugin-throttling": "^8.2.0",
-    "@octokit/types": "^13.5.0",
-    "@snyk/go-semver": "^1.4.0",
-    "execa": "^9.5.1",
-    "p-limit": "^6.1.0",
-    "vitest": "^2.0.0",
-    "zx": "^8.1.2"
-  },
-  "author": "@smartcontractkit",
-  "license": "MIT",
-  "repository": "https://github.com/smartcontractkit/.github"
-}
