--- apps/go-test-caching/project.json
@@ -1,64 +1,0 @@
-{
-  "name": "go-test-caching",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "projectType": "library",
-  "targets": {
-    "build": {
-      "dependsOn": ["build:internal"],
-      "command": "rimraf dist/package.json"
-    },
-    "typecheck": {
-      "executor": "nx:run-commands",
-      "options": {
-        "command": "tsc",
-        "cwd": "{projectRoot}"
-      }
-    },
-    "build:internal": {
-      "executor": "@nx/esbuild:esbuild",
-      "outputs": ["{options.outputPath}"],
-      "defaultConfiguration": "production",
-      "options": {
-        "format": ["cjs"],
-        "thirdParty": true,
-        "outputPath": "apps/go-test-caching/dist/",
-        "main": "apps/go-test-caching/src/index.ts",
-        "assets": [],
-        "outputFileName": "index.js",
-        "esbuildOptions": {
-          "outExtension": {
-            ".js": ".js"
-          }
-        },
-        "tsConfig": "tsconfig.json",
-        "platform": "node"
-      },
-      "configurations": {
-        "development": {
-          "minify": false
-        },
-        "production": {
-          "minify": false
-        }
-      }
-    },
-    "test": {
-      "executor": "nx:run-commands",
-      "options": {
-        "command": "vitest run",
-        "cwd": "{projectRoot}"
-      }
-    },
-    "lint": {
-      "executor": "@nx/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": [
-          "apps/go-test-caching/**/*.ts",
-          "apps/go-test-caching/package.json"
-        ]
-      }
-    }
-  },
-  "tags": []
-}
