--- apps/go-test-caching/src/index.ts
@@ -1,3 +1,0 @@
-import { run } from "./main.js";
-
-run();
