--- apps/go-test-caching/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "module": "commonjs",
-    "outDir": "./dist",
-    "resolveJsonModule": true,
-    "esModuleInterop": true
-  },
-  "include": ["src/**/*", "src/**/*.json"],
-  "exclude": ["src/**/*.spec.ts"]
-}
